cleaned up main, moved minishell_loop to own file
