+++ Makefile
@@ -6,7 +6,7 @@
 #    By: juramos <juramos@student.42madrid.com>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/09/15 16:43:25 by juramos           #+#    #+#              #
-#    Updated: 2024/04/17 10:29:33 by juramos          ###   ########.fr        #
+#    Updated: 2024/04/17 12:08:21 by juramos          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -32,6 +32,7 @@ SRC_FILES 	= 	main\
 				utils/clean\
 				utils/extra_libft\
 				utils/error\
+				utils/minishell_loop\
 				exec/arr_utils\
 				exec/exec\
 				exec/error_handlers\
